move additional dirs code to avoid circular dependency
